+++ src-docs/src/views/datagrid/datagrid.js
@@ -4,6 +4,8 @@ import React, {
   useEffect,
   useMemo,
   useState,
+  createContext,
+  useContext,
 } from 'react';
 import { fake } from 'faker';
 
@@ -17,15 +19,84 @@ import {
   EuiButtonIcon,
   EuiSpacer,
 } from '../../../../src/components/';
+const DataContext = createContext();
+
+const raw_data = [];
+
+for (let i = 1; i < 100; i++) {
+  const email = fake('{{internet.email}}');
+  const name = fake('{{name.lastName}}, {{name.firstName}}');
+  const suffix = fake('{{name.suffix}}');
+  raw_data.push({
+    name: {
+      formatted: `${name} ${suffix}`,
+      raw: name,
+    },
+    email: {
+      formatted: <EuiLink href="">{fake('{{internet.email}}')}</EuiLink>,
+      raw: email,
+    },
+    location: (
+      <Fragment>
+        {`${fake('{{address.city}}')}, `}
+        <EuiLink href="https://google.com">
+          {fake('{{address.country}}')}
+        </EuiLink>
+      </Fragment>
+    ),
+    date: fake('{{date.past}}'),
+    account: fake('{{finance.account}}'),
+    amount: fake('${{commerce.price}}'),
+    phone: fake('{{phone.phoneNumber}}'),
+    version: fake('{{system.semver}}'),
+  });
+}
 
 const columns = [
   {
     id: 'name',
     displayAsText: 'Name',
     defaultSortDirection: 'asc',
+    cellActions: [
+      ({ rowIndex, columnId, Component }) => {
+        const data = useContext(DataContext);
+        return (
+          <Component
+            onClick={() => alert(`Hi ${data[rowIndex][columnId].raw}`)}
+            iconType="heart"
+            aria-label={`Say hi to ${data[rowIndex][columnId].raw}!`}>
+            Say hi
+          </Component>
+        );
+      },
+      ({ rowIndex, columnId, Component }) => {
+        const data = useContext(DataContext);
+        return (
+          <Component
+            onClick={() => alert(`Bye ${data[rowIndex][columnId].raw}`)}
+            iconType="moon"
+            aria-label={`Say bye to ${data[rowIndex][columnId].raw}!`}>
+            Say bye
+          </Component>
+        );
+      },
+    ],
   },
   {
     id: 'email',
+    cellActions: [
+      ({ rowIndex, columnId, Component }) => {
+        const data = useContext(DataContext);
+        return (
+          <Component
+            onClick={() => alert(data[rowIndex][columnId].raw)}
+            iconType="email"
+            aria-label={`Send email to ${data[rowIndex][columnId].raw}`}>
+            Send email
+          </Component>
+        );
+      },
+    ],
   },
   {
     id: 'location',
@@ -46,6 +117,26 @@ const columns = [
         },
       ],
     },
+    cellActions: [
+      ({ rowIndex, columnId, Component, isExpanded }) => {
+        const data = useContext(DataContext);
+        const onClick = isExpanded
+          ? () =>
+              alert(`Sent money to ${data[rowIndex][columnId]} when expanded`)
+          : () =>
+              alert(
+                `Sent money to ${data[rowIndex][columnId]} when not expanded`
+              );
+        return (
+          <Component
+            onClick={onClick}
+            iconType="faceHappy"
+            aria-label={`Send money to ${data[rowIndex][columnId]}`}>
+            Send money
+          </Component>
+        );
+      },
+    ],
   },
   {
     id: 'date',
@@ -66,28 +157,6 @@ const columns = [
     actions: false,
   },
 ];
-
-const raw_data = [];
-
-for (let i = 1; i < 100; i++) {
-  raw_data.push({
-    name: fake('{{name.lastName}}, {{name.firstName}} {{name.suffix}}'),
-    email: <EuiLink href="">{fake('{{internet.email}}')}</EuiLink>,
-    location: (
-      <Fragment>
-        {`${fake('{{address.city}}')}, `}
-        <EuiLink href="https://google.com">
-          {fake('{{address.country}}')}
-        </EuiLink>
-      </Fragment>
-    ),
-    date: fake('{{date.past}}'),
-    account: fake('{{finance.account}}'),
-    amount: fake('${{commerce.price}}'),
-    phone: fake('{{phone.phoneNumber}}'),
-    version: fake('{{system.semver}}'),
-  });
-}
 
 const trailingControlColumns = [
   {
@@ -186,11 +255,12 @@ export default () => {
 
   const renderCellValue = useMemo(() => {
     return ({ rowIndex, columnId, setCellProps }) => {
+      const data = useContext(DataContext);
       useEffect(() => {
         if (columnId === 'amount') {
-          if (raw_data.hasOwnProperty(rowIndex)) {
+          if (data.hasOwnProperty(rowIndex)) {
             const numeric = parseFloat(
-              raw_data[rowIndex][columnId].match(/\d+\.\d+/)[0],
+              data[rowIndex][columnId].match(/\d+\.\d+/)[0],
               10
             );
             setCellProps({
@@ -200,33 +270,41 @@ export default () => {
             });
           }
         }
-      }, [rowIndex, columnId, setCellProps]);
+      }, [rowIndex, columnId, setCellProps, data]);
 
-      return raw_data.hasOwnProperty(rowIndex)
-        ? raw_data[rowIndex][columnId]
+      function getFormatted() {
+        return data[rowIndex][columnId].formatted
+          ? data[rowIndex][columnId].formatted
+          : data[rowIndex][columnId];
+      }
+
+      return data.hasOwnProperty(rowIndex)
+        ? getFormatted(rowIndex, columnId)
         : null;
     };
   }, []);
 
   return (
-    <EuiDataGrid
-      aria-label="Data grid demo"
-      columns={columns}
-      columnVisibility={{ visibleColumns, setVisibleColumns }}
-      trailingControlColumns={trailingControlColumns}
-      rowCount={raw_data.length}
-      renderCellValue={renderCellValue}
-      inMemory={{ level: 'sorting' }}
-      sorting={{ columns: sortingColumns, onSort }}
-      pagination={{
-        ...pagination,
-        pageSizeOptions: [10, 50, 100],
-        onChangeItemsPerPage: onChangeItemsPerPage,
-        onChangePage: onChangePage,
-      }}
-      onColumnResize={(eventData) => {
-        console.log(eventData);
-      }}
-    />
+    <DataContext.Provider value={raw_data}>
+      <EuiDataGrid
+        aria-label="Data grid demo"
+        columns={columns}
+        columnVisibility={{ visibleColumns, setVisibleColumns }}
+        trailingControlColumns={trailingControlColumns}
+        rowCount={raw_data.length}
+        renderCellValue={renderCellValue}
+        inMemory={{ level: 'sorting' }}
+        sorting={{ columns: sortingColumns, onSort }}
+        pagination={{
+          ...pagination,
+          pageSizeOptions: [10, 50, 100],
+          onChangeItemsPerPage: onChangeItemsPerPage,
+          onChangePage: onChangePage,
+        }}
+        onColumnResize={(eventData) => {
+          console.log(eventData);
+        }}
+      />
+    </DataContext.Provider>
   );
 };