+++ src-docs/src/views/datagrid/datagrid_styling_example.js
@@ -24,14 +24,19 @@ const dataGridControlsHtml = renderToHtml(DataGridControls);
 
 import DataGridColumnWidths from './column_widths';
 import DataGridColumnActions from './column_actions';
+import DataGridColumnCellActions from './column_cell_actions';
 const dataGridColumnWidthsSource = require('!!raw-loader!./column_widths');
 const dataGridColumnWidthsHtml = renderToHtml(DataGridColumnWidths);
 const dataGridColumnActionsSource = require('!!raw-loader!./column_actions');
 const dataGridColumnActionsHtml = renderToHtml(DataGridColumnActions);
+const dataGridColumnCellActionsSource = require('!!raw-loader!./column_cell_actions');
+const dataGridColumnCellActionsHtml = renderToHtml(DataGridColumnActions);
 
 import {
   EuiDataGridColumn,
   EuiDataGridColumnActions,
+  EuiDataGridColumnCellAction,
+  EuiDataGridColumnCellActionProps,
   EuiDataGridStyle,
   EuiDataGridToolBarVisibilityOptions,
 } from '!!prop-loader!../../../../src/components/datagrid/data_grid_types';
@@ -46,11 +51,11 @@ const gridSnippet = `<EuiDataGrid
   ]}
   // This can work as a shape.
   toolbarVisibility={{
-    showStyleSelector: false
-    showSortSelector: false
-    showFullScreenSelector: false
+    showStyleSelector: false,
+    showSortSelector: false,
+    showFullScreenSelector: false,
     // showColumnSelector also takes an object, check the prop docs.
-    showColumnSelector: false
+    showColumnSelector: false,
     additionalControls: (
       <Fragment>
         <EuiButtonEmpty
@@ -302,7 +307,6 @@ export const DataGridStylingExample = {
         </Fragment>
       ),
       components: { DataGridColumnActions },
-      snippet: widthsSnippet,
       props: {
         EuiDataGrid,
         EuiDataGridColumn,
@@ -310,6 +314,53 @@ export const DataGridStylingExample = {
         EuiListGroupItem,
       },
       demo: <DataGridColumnActions />,
+    },
+    {
+      source: [
+        {
+          type: GuideSectionTypes.JS,
+          code: dataGridColumnCellActionsSource,
+        },
+        {
+          type: GuideSectionTypes.HTML,
+          code: dataGridColumnCellActionsHtml,
+        },
+      ],
+      title: 'Column cell actions',
+      text: (
+        <Fragment>
+          <p>
+            On top of making a cell expandable, you can add more custom actions
+            by setting <EuiCode>cellActions</EuiCode>. This contains functions
+            called to render additional buttons in the cell and in the popover
+            when expanded. Behind the scenes those are treated as a React
+            components allowing hooks, context, and other React concepts to be
+            used. The functions receives an argument of type
+            <code>EuiDataGridColumnCellActionProps</code>. The icons of these
+            actions are displayed on mouse over, and also appear in the popover
+            when the cell is expanded. Note that once you&apos;ve defined the{' '}
+            <EuiCode>cellAction</EuiCode> property, the cell&apos;s
+            automatically expandable.
+          </p>
+          <p>
+            Below, the email and city columns provide 1{' '}
+            <EuiCode>cellAction</EuiCode> each, while the country column
+            provides 2 <EuiCode>cellAction</EuiCode>s. The city column shows
+            another action with different alert when it&apos;s clicked in the
+            popover.
+          </p>
+        </Fragment>
+      ),
+      components: { DataGridColumnCellActions },
+      props: {
+        EuiDataGrid,
+        EuiDataGridColumn,
+        EuiDataGridColumnActions,
+        EuiDataGridColumnCellAction,
+        EuiDataGridColumnCellActionProps,
+        EuiListGroupItem,
+      },
+      demo: <DataGridColumnCellActions />,
     },
   ],
 };